move batch to util/batch
