+++ client/src/util/batch/index.ts
@@ -1,0 +1,2 @@
+export * from "./batch";
+export * from "./interfaces";
